app/requirements.txt
Flask==1.0
google-cloud-storage
fireo